requirements.txt
python-dotenv
langchain_groq
streamlit
langchain
langchain-core
langchain-community
